# refactor: speed up IO